Fix the chart bug caused by the eager import @shopify/react-native-skia to the main bundle
